Fix some things. --toggle-analytics works
