+++ pros/ga/analytics.json
@@ -1,4 +1,4 @@
 {
-    "enabled" : true,
+    "enabled": "True",
     "ga_id": "GA-84548828-8"
 }
