+++ requirements.txt
@@ -12,4 +12,3 @@ scan-build==2.0.13
 sentry-sdk
 observable
 pypng==0.0.20
-pygamp
